requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
flask-mail==0.9.1
email_validator==1.1.2
PyGithub==1.55
beautifulsoup4==4.10.0
shortuuid==1.0.8
